Rafactor BooksForm and App
